Enhance logging

+++ src/lib/proxy.h
@@ -45,7 +45,7 @@ int proxy_init(proxy_ctx_t **proxy, http_req *req, http_res *res, http_status_ct
 
 int proxy_peek_response(proxy_ctx_t *proxy, http_res *res, http_status_ctx *ctx, http_status *custom_status, char *err_msg);
 
-int proxy_send(proxy_ctx_t *proxy, sock *client, unsigned long len_to_send, int flags);
+long proxy_send(proxy_ctx_t *proxy, sock *client, unsigned long len_to_send, int flags);
 
 int proxy_dump(proxy_ctx_t *proxy, char *buf, long len);
